feat: define vpc and runner ami
We have a requirement for build agents with improved speed and memory capacity than those offered by
the Github Actions infrastructure. This repository automates the creation of a VPC where EC2
instances can be launched and used as a runner for a job in Github Actions. The image for the runner
is also automated. Both configurations can be extended as need be, as and when new requirements
arise.