feat: webhook module
This module has been adapted from the `philips-labs/terraform-aws-github-runner` repository.

The webhook module defines a lambda function and an API gateway, along with the roles necessary for
the lambda to operate correctly. The API gateway provides an entry point to the lamba, and this is
where events from Github Actions will be posted to. A signature is sent along with the request to
verify it comes from a legitimate source.

The lambda function is implemented in TypeScript. A lot of the code was removed from the upstream
repository because it supported scenarios that were not necessary for us. The function receives
events from Github Actions, and if it's a `workflow_event`, it will put it on an SQS queue, which is
going to be read by another lambda.