feat: remove webhook module and lambda
The webhook module, which was adapted from the philips-labs repository, is removed to make way for a
simpler setup.

In the new setup, we can directly launch an EC2 instance, so the SQS queue is not necessary. We will
also host two Lambda functions packaged as containers, so two repositories are created in our
private container registry. This registry is something that's automatically available for any AWS
account, so it doesn't need to be created with Terraform.