feat: lambda function for launching ec2 instance
The Github app for the self-hosted runner is installed at the organisation level. The app allows you
to provide a webhook where it will post a `workflow_job` event. This webhook hits an AWS Lambda
function, which then launches an EC2 instance.

The function is implemented in Python. It gets deployed using the AWS SAM CLI, which uses
CloudFormation, and there is a template defined in this repository. There are details on the
deployment process in the README.

An EC2 instance will only be launched when the `action` for the `workflow_job` is `queued` and the
label is `self-hosted` (this is the value of `runs-on` in the workflow definition). This is
important because lots of these events are received and we only want an instance launched for
self-hosted jobs. Due to the webhook being publicly available, we also protect against DDoS by
verifying the signature that is sent by Github.