refactor: reorganise lambda function
Initially I thought I would have two functions that would be deployed independently, but I realised
I only need one function. When the job completes, an event with a `completed` action will be
received. This can be used to remove any idle runners and the corresponding EC2 instance.

So the code from the `gha-runner-create-instance` sub directory moves up a level and the module is
renamed from `create_instance` to `manage_runners`.

I've also generated a new private key for the Github app as I was having some problems with the
configuration.