feat: remove idle runners for completed job
When the job running on the self-hosted runner is completed, Github Actions will send another
`workflow_job` event, this time with the `action` `completed`. We can use the same Lambda function
to remove the EC2 instance that was used for the job. One of these `workflow_job` events will be
received for every completed job, so as with the `queued` event, we only process the events that
have a `self-hosted` label.

First the runner is removed from Github Actions, then the corresponding EC2 instance is deleted. The
runner and the instance are tied together with the private IP address of the instance, which is the
name assigned to the runner.

Since we only need one Lambda function, this also removes the previous two Docker repositories and
replaces them with a single repository.