feat: add extra disk to ec2 instance
The main disk on the instance is nearly full, which prevents safe_network from being able to build
completely. The dependency graph uses space and Cargo also makes use of temporary space during its
build process.

The Packer template is updated to instruct the AMI to attach an extra 50gb disk to the instance
that gets launched, and the user data script gets extended to format and mount the disk. We create
directories for tmp, `CARGO_HOME` and the Github runner, where the workspace and code will be
checked out. There is a little loop in the script to wait for the additional disk to become
available, since sometimes it's not attached by the time this script runs.

In any actions job that uses the self-hosted runner we should set `TMPDIR` and `CARGO_HOME` to the
locations provided here.