feat: extend ami with cargo and musl
Several packages are provided that allow the machine to perform a complete build of the
`safe_network` repository, including test code.

The Cargo installation symlinks all the binaries to a system-wide location. This is because I ran
into some problems with the runner process not having a full login shell environment and I had to
end up having a `source ~/.cargo/env` command in every `shell` step that used Cargo. We can just
symlink the binaries to `/usr/local/bin` to avoid this.