docs: explain the self hosted build agent setup
After coming back to this repository two years later and trying to figure out how to get it running
again, I decided to document anything I found lacking. It turns out there wasn't really much of a
description of the different parts in the infrastructure and how they fit together.